Update GSI to run with WRF
Updates mainly focus on undoing the work in commit 460d63f, which switched the 2-m temperature analysis variable from potential to sensible temperature. The intermediate binary file used by GSI (sigfXX) currently only contains 2-m potential temperature when using WRF, so asking for 2-m sensible temperature raises an error. In the future, we should add 2-m sensible temperature as a field in sigfXX, but this will work for now as a hot fix.